+++ lib/plaid/products/sandbox.rb
@@ -44,6 +44,24 @@ module Plaid
       # Public: The Boolean webhook fired success flag.
       property :webhook_fired
     end
+
+    # Public: Sets the verification status for an item
+    # created via automated microdeposits
+    #
+    # Does a POST /sandbox/item/set_verification_status call.
+    #
+    # access_token - access_token of the item
+    # account_id - id of the account to verify
+    # verification_status - status to set
+    #
+    # Returns a Models::BaseResponse object.
+    def set_verification_status(access_token, account_id, verification_status)
+      post_with_auth '/sandbox/item/set_verification_status',
+                     Models::BaseResponse,
+                     access_token: access_token,
+                     account_id: account_id,
+                     verification_status: verification_status
+    end
   end
 
   # Public: Class used to call the SandboxPublicToken sub-product
@@ -54,10 +72,12 @@ module Plaid
     # to generate a public_token given an institution and list of
     # products.
     #
-    # institution_id   - Specific institution id to generate token for.
-    # initial_products - Products for which generated token is valid for.
-    # webhook          - webhook to associate with the item (optional).
-    # options          - Additional options to merge into API request.
+    # institution_id    - Specific institution id to generate token for.
+    # initial_products  - Products for which generated token is valid for.
+    # webhook           - webhook to associate with the item (optional).
+    # override_username - Specific test credential username to use
+    # override_password - Specific test credential password to use
+    # options           - Additional options to merge into API request.
     #
     # Returns a SandboxCreateResponse object with a public token and item id.
     def create(institution_id:,
@@ -65,6 +85,8 @@ module Plaid
                transactions_start_date: nil,
                transactions_end_date: nil,
                webhook: nil,
+               override_username: nil,
+               override_password: nil,
                options: nil)
 
       options_payload = {}
@@ -72,6 +94,12 @@ module Plaid
       txn_options = transaction_options transactions_start_date,
                                         transactions_end_date
       options_payload[:transactions] = txn_options if txn_options != {}
+      unless override_username.nil?
+        options_payload[:override_username] = override_username
+      end
+      unless override_password.nil?
+        options_payload[:override_password] = override_password
+      end
       options_payload = options_payload.merge(options) unless options.nil?
 
       post_with_public_key 'sandbox/public_token/create',