+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# 10.0.0
+
+- Add support for optional user fields for `/item/add_token/create` endpoint ([278](https://github.com/plaid/plaid-ruby/pull/278))
+- Add support for credit card liabilities ([286](https://github.com/plaid/plaid-ruby/pull/286))
+
+BREAKING CHANGES:
+
+- `RATE_LIMIT_EXCEEDED` Plaid error types will be correctly mapped to `RateLimitExceededError` ([285](https://github.com/plaid/plaid-ruby/pull/285))
+- `INSTITUTION_ERROR` Plaid error types will be correctly mapped to `InstitutionError` ([275](https://github.com/plaid/plaid-ruby/pull/275))
+- Enable payment recipient to be created with BACS ([288](https://github.com/plaid/plaid-ruby/pull/288))
+
+# 9.0.0
+
+- Adds support for `/sandbox/item/set_verification_status`
+- Removes support for deprecated `/item/access_token/update_version` endpoint
+
+BREAKING CHANGES:
+
+- Removes `client.item.update_version`
+
+# 8.9.0 16-Apr-2020
+
+- Add client_user_id to Item add token endpoint ([#274](https://github.com/plaid/plaid-ruby/pull/274))
+
 # 8.8.0 02-Apr-2020
 
 - Add support for Item add token endpoint (BETA) ([#269](https://github.com/plaid/plaid-ruby/pull/269))